backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic
pydantic-settings
python-jose
python-dotenv
pandas
numpy
scikit-learn
xgboost
joblib
google-auth
httpx
openai
email-validator
requests
